Added Maven pom.xml
Added the Maven pom.xml containing only the project's dependencies.
For the Oracle JDBC driver i needed to create a separate command stored
in maven_install_oracle_jdbc_driver.bat. This allows you to install the
driver as a local artifact usable by Maven. The reason we had to do this
is because it's not officialy available on http://search.maven.org/ or
other dependency storage websites.